requirements.txt
tensorflow
flask
fastapi
uvicorn
requests
pillow
